requirements.txt
appdirs==1.4.4
APScheduler==3.7.0
astroid==2.4.2
Babel==2.8.0
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
crispy-tailwind==0.3.0
decorator==4.4.2
distlib==0.3.1
Django==2.1.8
django-admin-shortcuts==2.0.0
django-countries==6.1.2
django-crispy-forms==1.11.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-phonenumber-field==4.0.0
django-simple-geo @ file:///C:/Users/HP/Desktop/pyprog/django_simple_geo-0.4.3-py2.py3-none-any.whl
django-tinymce==3.2.0
djangorestframework==3.11.1
docutils==0.16
filelock==3.0.12
get==0.0.0
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.2
jupyter-client==6.1.7
jupyter-core==4.6.3
Kivy==1.11.1
kivy-deps.glew==0.2.0
kivy-deps.sdl2==0.2.0
Kivy-Garden==0.1.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.18
mysqlclient==1.4.5
parso==0.7.1
phonenumbers==8.12.5
pickleshare==0.7.5
Pillow==7.2.0
post==0.0.0
prompt-toolkit==3.0.7
protobuf==3.12.2
psycopg2==2.8.3
pycoingecko @ file:///C:/Users/HP/Desktop/pyprog/pycoingecko-1.3.0-py3-none-any.whl
Pygments==2.6.1
pylint==2.5.3
pylint-django==2.3.0
pylint-plugin-utils==0.6
pypiwin32==223
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pyzmq==19.0.2
query-string==0.0.0
request==0.0.0
requests==2.24.0
six==1.15.0
toml==0.10.1
tornado==6.0.4
traitlets==5.0.4
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.25.10
virtualenv==20.0.26
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.1.0
